+++ Assets/LDtkUnity/Editor/Builders/LDtkBuilderEntity.cs
@@ -203,7 +203,7 @@ namespace LDtkUnity.Editor
                 Vector2 pivotOffset = LDtkCoordConverter.EntityPivotOffset(entityDef.UnityPivot, size);
                 Vector3 middleCenter = gameObject.transform.position + (Vector3)pivotOffset;
                 
-                LDtkFieldDrawerData data = new LDtkFieldDrawerData(fields, smartColor, displayMode, fieldInstance.Identifier, gridSize, middleCenter);
+                LDtkFieldDrawerData data = new LDtkFieldDrawerData(fields, smartColor, displayMode, fieldInstance.Identifier, gridSize, Importer.PixelsPerUnit, middleCenter);
                 drawerComponent.AddReference(data);
             }
         }