+++ Assets/LDtkUnity/Editor/CustomEditor/SceneDrawer/ComponentDrawers/LDtkSceneDrawerField.cs
@@ -1,34 +1,11 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace LDtkUnity.Editor
 {
     internal static class LDtkSceneDrawerField
     {
-        public static void Draw(List<LDtkEntityDrawerComponent> entityComponents)
-        {
-            List<LDtkFieldDrawerData> datas = entityComponents.SelectMany(component => component.FieldDrawers).ToList();
-
-            foreach (LDtkFieldDrawerData data in datas)
-            {
-                ProcessData(data);
-            }
-        }
-        private static void ProcessData(LDtkFieldDrawerData data)
-        {
-            if (!data.Enabled)
-            {
-                return;
-            }
-            
-            Handles.color = data.SmartColor;
-            ILDtkHandleDrawer drawer = DrawField(data);
-            drawer?.OnDrawHandles();
-        }
-
-        private static ILDtkHandleDrawer DrawField(LDtkFieldDrawerData data)
+        public static ILDtkHandleDrawer DrawField(LDtkFieldDrawerData data)
         {
             if (data.Fields == null)
             {
@@ -48,14 +25,14 @@ namespace LDtkUnity.Editor
                     return new LDtkFieldDrawerValue(data.Fields.transform.position + Vector3.up, data.Identifier);
                     
                 case EditorDisplayMode.EntityTile: 
-                    //this is actually handled in the entity drawer, not here
+                    //this is actually handled in the entity drawer, not here. The only reason why is because we want the entity identifier to render a little bit lower if the icon exists. also certain responsibilities handled there
                     break;
 
                 case EditorDisplayMode.PointPath:
                 case EditorDisplayMode.PointStar:
                 case EditorDisplayMode.PointPathLoop:
                 case EditorDisplayMode.Points:
-                    return new LDtkFieldDrawerPoints(data.Fields, data.Identifier, data.FieldMode, data.MiddleCenter, data.GridSize);
+                    return new LDtkFieldDrawerPoints(data.Fields, data.Identifier, data.FieldMode, data.MiddleCenter, data.PixelsPerUnit);
                     
                 case EditorDisplayMode.RadiusPx:
                 case EditorDisplayMode.RadiusGrid:
@@ -68,7 +45,7 @@ namespace LDtkUnity.Editor
                     
                 case EditorDisplayMode.RefLinkBetweenCenters:
                 case EditorDisplayMode.RefLinkBetweenPivots:
-                    return new LDtkFieldDrawerEntityRef(data.Fields, data.Identifier, data.FieldMode, data.GridSize, data.MiddleCenter, data.SmartColor);
+                    return new LDtkFieldDrawerEntityRef(data.Fields, data.Identifier, data.FieldMode, data.PixelsPerUnit, data.MiddleCenter, data.SmartColor);
 
                 default:
                     return null;